chore(rust): replace previous `ockam_command` with `ockam-cli-demo`
