+++ implementations/rust/ockam/ockam_command/Cargo.toml
@@ -4,26 +4,43 @@ version = "0.11.0"
 edition = "2021"
 license = "Apache-2.0"
 publish = false
-rust-version = "1.56.0"
+rust-version = "1.58.1"
+authors = ["Ockam Developers"]
+homepage = "https://github.com/ockam-network/ockam"
+repository = "https://github.com/ockam-network/ockam/implementations/rust/ockam/ockam_command"
+description = ""
 
 [[bin]]
-name = "ockam-cli"
+# You may be wondering "Why are the tests and docs disabled?". The long and
+# short of it is: To avoid certain bugs in `rustdoc`, `cargo`, and other tools
+# which have cases (even edge-cases) where they confuse the library (in
+# `../ockam`) and the `ockam` binary (in `./src/bin/ockam.rs`). I won't
+# enumerate them here, but an example: `rustdoc` will try to place the docs for
+# both of these in the same path, without realizing it, which may result in one
+# overwriting the other)
+#
+# Anyway a result, we disable them for the binary crate, which is just a single
+# file (`src/bin/ockam.rs`) which contains a single function call into
+# `ockam_command` itself. This way, while there are no tests/docs for the binary
+# crate, it doesn't matter, since its just a shim for the `ockam_command` crate
+# anyway (which does not have its tests/docs disabled)
+name = "ockam"
+doc = false
+test = false
 
 [dependencies]
-bunt = "0.2.6"
-clap = { version = "3.0.4", features = ["derive"] }
-comfy-table = "5.0.0"
-config = { version = "0.11.0" }
-ctrlc = "3.2.1"
-dotenv = "0.15.0"
-env_logger = "0.9.0"
-exitcode = "1.1.2"
-human-panic = "1.0.3"
-indicatif = "0.16.2"
-log = "0.4.14"
-ockam = { version = "^0.51.0", path = "../ockam" }
-ockam_core = { path = "../ockam_core" }
-ockam_vault = { path = "../ockam_vault" }
-serde = "1.0.130"
-ssh-key = "0.2.0"
-thiserror = "1.0.30"
+anyhow = "1"
+dirs = "4.0.0"
+clap = { version = "3", features = ["derive", "cargo"] }
+tracing = { version = "0.1.31", features = ["attributes"] }
+tracing-subscriber = "0.3.9"
+tokio = { version="1", features = ["full"] }
+async-trait = "0.1"
+hex = "0.4"
+serde_json = "1"
+serde = { version = "1", features = ["derive"] }
+rand = "0.8"
+
+ockam = { path = "../ockam", version = "0.51.0", features = ["no_init_tracing", "software_vault"] }
+ockam_vault = { path = "../ockam_vault", version = "0.43.0", features = ["storage"] }
+ockam_core = { path = "../ockam_core", version = "0.50.0" }
