--- implementations/rust/ockam/ockam_command/ockam_secrets.toml
@@ -1,1 +1,0 @@
-SEKRET = "yes"
