+++ implementations/rust/ockam/ockam_command/src/cmd/outlet.rs
