--- implementations/rust/ockam/ockam_command/src/command.rs
@@ -1,4 +1,0 @@
-pub mod channel;
-pub mod channel_listen;
-pub mod inlet;
-pub mod outlet;
