+++ implementations/rust/ockam/ockam_command/src/lib.rs
@@ -1,44 +1,205 @@
-pub mod command;
-pub mod config;
-pub mod service;
-pub mod spinner;
+//! This library exists only to be used by the `ockam` CLI (in
+//! `./bin/ockam.rs`).
+//!
+//! It should only publically expose a [single item](`run_main`) which should
+//! called from `fn main()` of `bin/ockam.rs` as the only thing it does.
 
-use std::io::Error;
-pub use thiserror::Error;
+use self::args::*;
+use anyhow::{Context, Result};
+use clap::Parser;
+use identity::load_identity;
+use ockam::identity::IdentityIdentifier;
+use std::collections::BTreeSet;
+use storage::{ensure_identity_exists, get_ockam_dir};
+use tracing_subscriber::{filter::LevelFilter, fmt, EnvFilter};
 
-#[derive(Error, Debug)]
-pub enum AppError {
-    #[error("unknown error")]
-    Unknown,
-    #[error("invalid command")]
-    InvalidCommand,
-    #[error("invalid argument")]
-    InvalidArgument,
+pub(crate) type OckamVault = ockam::vault::VaultMutex<ockam_vault::SoftwareVault>;
 
-    #[error("i/o error")]
-    IoError(std::io::Error),
-
-    #[error("ssh key error")]
-    SshKey(ssh_key::Error),
-
-    #[error("ockam error")]
-    Ockam(ockam::Error),
+pub(crate) mod args;
+pub(crate) mod identity;
+pub(crate) mod storage;
+pub(crate) mod cmd {
+    pub(crate) mod identity;
+    pub(crate) mod inlet;
+    pub(crate) mod outlet;
 }
 
-impl From<ockam::Error> for AppError {
-    fn from(ockam_error: ockam::Error) -> Self {
-        AppError::Ockam(ockam_error)
+// This should be this library's only public function.
+pub fn run_main() {
+    let args = CliArgs::parse();
+    let verbose = args.verbose;
+    init_logging(verbose);
+    tracing::debug!("Parsed arguments (outlet) {:?}", args);
+    // Note: We don't force all commands to start the node.
+    match args.command {
+        args::Command::CreateIdentity(arg) => node_subcommand(verbose > 0, arg, cmd::identity::run),
+        args::Command::CreateInlet(arg) => node_subcommand(verbose > 0, arg, cmd::inlet::run),
+        args::Command::CreateOutlet(arg) => node_subcommand(verbose > 0, arg, cmd::outlet::run),
+        args::Command::AddTrustedIdentity(arg) => exit_with_result(verbose > 0, add_trusted(arg)),
+        args::Command::PrintIdentity => exit_with_result(verbose > 0, print_identity()),
+        args::Command::PrintPath => exit_with_result(verbose > 0, print_ockam_dir()),
     }
 }
 
-impl From<std::io::Error> for AppError {
-    fn from(io_error: Error) -> Self {
-        AppError::IoError(io_error)
+fn print_identity() -> anyhow::Result<()> {
+    ensure_identity_exists(false)?;
+    let dir = get_ockam_dir()?;
+    let identity = load_identity(&dir.join("identity.json"))?;
+    println!("{}", identity.id.key_id());
+    Ok(())
+}
+
+fn print_ockam_dir() -> anyhow::Result<()> {
+    match get_ockam_dir() {
+        Ok(path) => {
+            // We'd rather panic than print a lossy (and thus possibly wrong)
+            // path. But `get_ockam_dir()` checks this.
+            println!("{}", path.to_str().expect("bug in `get_ockam_dir`"));
+            Ok(())
+        }
+        Err(e) => {
+            eprintln!(
+                "Failed to locate the ockam directory (or it was invalid). \
+                Hint: try providing `$OCKAM_DIR` explicitly, or changing the \
+                value of `$OCKAM_DIR` if it is already set in your environment.",
+            );
+            Err(e)
+        }
     }
 }
 
-impl From<ssh_key::Error> for AppError {
-    fn from(ssh_error: ssh_key::Error) -> Self {
-        AppError::SshKey(ssh_error)
+fn add_trusted(arg: AddTrustedIdentityOpts) -> anyhow::Result<()> {
+    // Parse args before we start complaining about the directory.
+    let to_trust = crate::identity::parse_identities(&arg.to_trust)?;
+    ensure_identity_exists(false)?;
+    let ockam_dir = get_ockam_dir()?;
+    let trusted_file = ockam_dir.join("trusted");
+    if to_trust.is_empty() && !arg.only {
+        eprintln!(
+            "No change to {} needed, no identities were \
+            provided (and `--only` is not in use).",
+            trusted_file.display(),
+        );
+        return Ok(());
+    }
+    let existing = if trusted_file.exists() && !arg.only {
+        crate::identity::read_trusted_idents_from_file(&trusted_file)?
+    } else {
+        vec![]
+    };
+    let need = to_trust
+        .into_iter()
+        .filter(|id| !existing.contains(id))
+        .collect::<Vec<_>>();
+    if need.is_empty() && !arg.only {
+        eprintln!(
+            "No change to {} needed, all identities already \
+            trusted (and `--only` is not in use).",
+            trusted_file.display(),
+        );
+        return Ok(());
+    }
+    let all: Vec<_> = existing.iter().chain(need.iter()).cloned().collect();
+    let all_dedup = all
+        .iter()
+        .cloned()
+        .collect::<BTreeSet<IdentityIdentifier>>();
+    // Keep user-provided order if no duplicates.
+    let idents_to_write = if all_dedup.len() == all.len() {
+        all
+    } else {
+        all_dedup.into_iter().collect::<Vec<_>>()
+    };
+    if idents_to_write == existing && !arg.only {
+        eprintln!(
+            "No change to {} needed. New and old trusted lists \
+            would be identical (and `--only` is not in use).",
+            trusted_file.display(),
+        );
+        return Ok(());
+    }
+    let strings_to_write = idents_to_write
+        .into_iter()
+        .map(|s| s.key_id().clone())
+        .collect::<Vec<String>>();
+    let new_contents = strings_to_write.join("\n");
+
+    crate::storage::write(&trusted_file, new_contents.as_bytes()).with_context(|| {
+        format!(
+            "Writing updated list of trusted identities to {:?}",
+            trusted_file
+        )
+    })?;
+    eprintln!(
+        "Wrote updated list to {}, containing {} identities.",
+        trusted_file.display(),
+        strings_to_write.len(),
+    );
+    Ok(())
+}
+
+fn node_subcommand<A, F, Fut>(verbose: bool, arg: A, f: F)
+where
+    A: Send + Sync + 'static,
+    F: FnOnce(A, ockam::Context) -> Fut + Send + Sync + 'static,
+    Fut: core::future::Future<Output = anyhow::Result<()>> + Send + 'static,
+{
+    let (ctx, mut executor) = ockam::start_node();
+    let res = executor.execute(async move {
+        if let Err(e) = f(arg, ctx).await {
+            print_error_and_exit(verbose, e);
+        }
+    });
+    if let Err(e) = res {
+        eprintln!(
+            "Ockam node failed at last minute. TODO: find out something \
+            smarter to do if this happens: {:?}",
+            e,
+        );
+    }
+}
+
+fn print_error_and_exit(v: bool, e: anyhow::Error) -> ! {
+    tracing::trace!("Exiting with error {:?}", e);
+    eprintln!("Error: {}", message(v, &e));
+    for cause in e.chain().skip(1) {
+        eprintln!("- caused by: {}", message(v, cause));
+    }
+    std::process::exit(1);
+}
+
+fn exit_with_result(verbose: bool, result: Result<()>) -> ! {
+    if let Err(e) = result {
+        print_error_and_exit(verbose, e);
+    } else {
+        std::process::exit(0);
+    }
+}
+
+fn message(verbose: bool, e: impl std::fmt::Display + std::fmt::Debug) -> String {
+    if verbose {
+        format!("{:?}", e)
+    } else {
+        format!("{}", e)
+    }
+}
+
+fn init_logging(verbose: u8) {
+    let filter = EnvFilter::try_from_env("OCKAM_LOG")
+        .unwrap_or_else(|_| {
+            if verbose == 0 {
+                EnvFilter::default().add_directive("ockam_node=warn".parse().unwrap())
+            } else {
+                EnvFilter::default()
+            }
+        })
+        .add_directive(match verbose + 1 {
+            0 => LevelFilter::WARN.into(),
+            1 => LevelFilter::INFO.into(),
+            2 => LevelFilter::DEBUG.into(),
+            _ => LevelFilter::TRACE.into(),
+        });
+    if fmt().with_env_filter(filter).try_init().is_err() {
+        tracing::warn!("Failed to initialise logging.");
     }
 }
