--- implementations/rust/ockam/ockam_command/src/service.rs
@@ -1,1 +1,0 @@
-pub mod echoer;
