--- integrations/suborbital/demo/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "ockam-cli-demo"
-version = "0.1.0"
-authors = ["Ockam Developers"]
-edition = "2021"
-license = "Apache-2.0"
-publish = false
-rust-version = "1.58.0"
-
-[[bin]]
-name = "ockam"
-doc = false
-test = false
-
-[dependencies]
-anyhow = "1"
-dirs = "4.0.0"
-clap = { version = "3", features = ["derive", "cargo"] }
-tracing = { version = "0.1.31", features = ["attributes"] }
-tracing-subscriber = "0.3.9"
-tokio = {version="1", features = ["full"]}
-async-trait = "0.1"
-hex = "0.4"
-serde_json = "1"
-serde = { version = "1", features = ["derive"] }
-rand = "0.8"
-subtle = "2"
-zeroize = {version="1", features=["zeroize_derive"]}
-[dependencies.ockam]
-path = "../../../implementations/rust/ockam/ockam"
-features = ["no_init_tracing"]
-[dependencies.ockam_vault]
-path = "../../../implementations/rust/ockam/ockam_vault"
-[dependencies.ockam_core]
-path = "../../../implementations/rust/ockam/ockam_core"
