--- integrations/suborbital/demo/src/bin/ockam.rs
@@ -1,5 +1,0 @@
-// This file exists to hack around some limitations cargo workspaces have around binary names.
-
-fn main() {
-    ockam_cli_demo::run_main()
-}
